price fixed

+++ deliveboo/database/migrations/2021_02_26_155956_create_orders_table.php
@@ -20,7 +20,7 @@ class CreateOrdersTable extends Migration
             $table -> string('name',60) -> nullable();
             $table -> string('lastname',60) -> nullable();
             $table -> string('address',60) -> nullable();
-            $table -> integer('tot_price') -> nullable();
+            $table -> float('tot_price') -> nullable();
             $table -> integer('month') -> nullable();
             $table -> string('phone',20) -> nullable();
             $table -> boolean('status') -> nullable();
